# modules/headless-counter/headless-counter-impl/rest-openapi.yaml
info:
    description:
        "HeadlessCounter REST API"
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: "HeadlessCounter"
    version: v1.0
openapi: 3.0.1
paths:
    "/counter/getnames":
        get:
            responses:
                200:
                    description: h
                    content:
                        application/json:
                            schema:
                                type: string
                        application/xml:
                            schema:
                                type: string
            summary: get names
            tags:
                - "Counter"
    "/counter/increment":
        get:
            responses:
                200:
                    description: h
                    content:
                        application/json:
                            schema:
                                format: int64
                                type: integer
                        application/xml:
                            schema:
                                format: int64
                                type: integer
            summary: impl for counter-imcrement
            tags:
                - "Counter"
    "/counter/increment/{name}":
        get:
            parameters:
                - in: path
                  name: name
                  required: true
                  schema:
                      type: string
            responses:
                200:
                    description: h
                    content:
                        application/json:
                            schema:
                                format: int64
                                type: integer
                        application/xml:
                            schema:
                                format: int64
                                type: integer
            summary: impl for counter-imcrement
            tags:
                - "Counter"
    "/counter/increment/{name}/{size}":
        get:
            parameters:
                - in: path
                  name: name
                  required: true
                  schema:
                      type: string
                - in: path
                  name: size
                  required: true
                  schema:
                      format: int32
                      type: integer
            responses:
                200:
                    description: hi
                    content:
                        application/json:
                            schema:
                                format: int64
                                type: integer
                        application/xml:
                            schema:
                                format: int64
                                type: integer
            summary: impl for counter-imcrement
            tags:
                - "Counter"
    "/counter/rename/{oldName}/{newName}":
        get:
            parameters:
                - in: path
                  name: oldName
                  required: true
                  schema:
                      type: string
                - in: path
                  name: newName
                  required: true
                  schema:
                      type: string
            responses:
                204:
                    description: h
                    content: {}
            summary: rename
            tags:
                - "Counter"
    "/counter/reset/{name}":
        get:
            parameters:
                - in: path
                  name: name
                  required: true
                  schema:
                      type: string
            responses:
                204:
                    content: {}
                    description:
                        tnx
            summary: impl for reset
            tags:
                - "Counter"
    "/counter/reset/{name}/{size}":
        get:
            parameters:
                - in: path
                  name: name
                  required: true
                  schema:
                      type: string
                - in: path
                  name: size
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                204:
                    content: {}
                    description:
                        tnx
            summary: impl for reset
            tags:
                - "Counter"
    "/test/{id}":
        get:
          description: its a test
          parameters:
            - in: path
              name: id
              required: true
              schema:
                type: integer
                format: int64
          responses:
              200:
                  description: hi
                  content:
                      application/json:
                          schema:
                            type: string
                      application/xml:
                          schema:
                            type: string
          tags:
            - "Counter"